convert data to FloatTensor

+++ visual-hyperplane/train_gan.py
@@ -261,7 +261,7 @@ for iteration in tqdm(xrange(args.iterations)):
 
     for iter_d in xrange(args.critic_iterations):
         # Real data
-        real_data = data_iter.next()
+        real_data = data_iter.next().type(torch.FloatTensor)
         # real_data = real_data + 0.1 * torch.rand(real_data.size())
         if args.use_cuda:
             real_data = real_data.cuda()
